Split Zvbb into 2-stage functional unit
